LangFuse and FastApi/requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
langfuse
fastapi
PyJWT
bcrypt
pgvector
uvicorn
fastapi-login
email-validator
async-generator
psycopg2
python-multipart